Poopy Pfusch

+++ Assets/Scripts/Classes/Abilities/Basic.cs
@@ -7,7 +7,7 @@ public class Basic : Attacks {
 
     private static int instanceCount = 0;
 
-    public Basic(GameObject caster, float damage, float castTime, float delay, float duration, float cooldown, float range, int targets, uint spellDir) : base(
+    public Basic(GameObject caster, int damage, float castTime, float delay, float duration, float cooldown, float range, int targets, uint spellDir) : base(
         caster, 0, "basic", "meele", targets, damage, castTime, delay, duration, cooldown, range, spellDir) { instanceCount++; }
 
     public override void StartSpell()
