+++ Assets/Scripts/Classes/Abilities/Virgo.cs
@@ -10,7 +10,7 @@ public class Virgo : Attacks {
     private float knockBackDirection;
     private float stunTime;
 
-    public Virgo(GameObject caster, float damage, float castTime, float delay, float duration, float cooldown, float range, int targets, uint spellDir, float dashStrength, float dashTime, float knockBackStrength, float knockBackTime, float stunTime) : base(
+    public Virgo(GameObject caster, int damage, float castTime, float delay, float duration, float cooldown, float range, int targets, uint spellDir, float dashStrength, float dashTime, float knockBackStrength, float knockBackTime, float stunTime) : base(
         caster, 6, "virgo", "utility ", targets, damage, castTime, delay, duration, cooldown, range, spellDir)
     {
         this.knockBackStrength = knockBackStrength;
